+++ CHANGES.md
@@ -157,7 +157,7 @@ To be released.
      -  Added `StateCompleters<T>` static class.
      -  Added `FungibleAssetStateCompleters<T>` static class.
      -  Added `Swarm<T>.GetTrustedStateCompleterAsync()` method.
- -  Added `IRenderer` interface.  [[#959], [#963]]
+ -  Added `IRenderer<T>` interface.  [[#959], [#963]]
  -  Added `BlockChain<T>.Renderers` property.  [[#959], [#963]]
  -  Added `BlockChain<T>.Render` property.  [[#946]]
  -  Added `Reorged` event on `BlockChain<T>`.  [[#945]]